Fixed ordering of enabled features in gwtoolbox settings

+++ GWToolboxdll/Modules/ToolboxSettings.cpp
@@ -186,7 +186,7 @@ void ToolboxSettings::DrawSettingInternal() {
         {"Vanquish counter",&use_vanquish}
     };
     ImGui::Columns(static_cast<int>(cols), "global_enable_cols", false);
-    size_t items_per_col = (size_t)ceil(features.size() / cols);
+    size_t items_per_col = (size_t)ceil(features.size() / static_cast<float>(cols));
     size_t col_count = 0;
     for (auto feature : features) {
         ImGui::Checkbox(feature.first, feature.second);
